pyproject.toml
[tool.poetry]
name = "workhorse"
version = "0.4.0"
description = "Manage GitHub repositories in bulk"
authors = ["Dave Gaeddert <dave.gaeddert@dropseed.io>"]
license = "MIT"

[tool.poetry.scripts]
workhorse = "workhorse.cli:cli"

[tool.poetry.dependencies]
python = "^3.6"
marshmallow = "^3.8.0"
jinja2 = "^2.11.2"
requests = "^2.24.0"
click = "^7.1.2"
pyyaml = "^5.3.1"
cached_property = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
ipdb = "^0.13.3"
bandit = "^1.6.2"
combine = "^2.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
